requirements.txt
alembic==1.6.5
aniso8601==9.0.1
attrs==22.2.0
awscrt==0.19.2
black==22.10.0
boto3==1.17.24
botocore==1.20.112
cachetools==5.3.1
certifi==2021.10.8
cffi==1.15.0
chardet==5.2.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.6
coverage==7.2.1
distlib==0.3.7
e==1.4.5
et-xmlfile==1.1.0
exceptiongroup==1.1.0
filelock==3.12.4
flash==1.0.3
Flask==2.0.1
Flask-RESTful==0.3.9
greenlet==1.1.2
httplib2==0.21.0
hypothesis==6.68.2
idna==3.3
iniconfig==2.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
Mako==1.1.6
MarkupSafe==2.1.3
mock==5.0.1
moto==4.1.3
mypy-extensions==0.4.3
numpy==1.23.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.1
packaging==23.1
pathspec==0.10.1
platformdirs==3.10.0
pluggy==1.3.0
protobuf==4.22.0
psycopg2==2.9.7
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyathena==2.10.0
pycparser==2.21
pyparsing==3.0.9
pyproject-api==1.6.1
pytest==7.2.1
pytest-cov==4.0.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==0.17.0
python-editor==1.0.4
pytz==2021.3
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.22.0
rsa==4.9
s3transfer==0.3.7
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.19
tenacity==8.2.1
toml==0.10.2
tomli==2.0.1
tox==4.11.3
types-toml==0.10.8.5
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.24.5
waitress==2.0.0
Werkzeug==2.2.2
xmltodict==0.13.0
